Adds Node setup and dependency installation to deploy workflow
This commit introduces two changes to the Github Actions deploy workflow and Dockerfile. Specifically, it adds a new step to the workflow for setting up a Node.js environment and installing project dependencies, meeting the requirement of Node version 18. Additionally, it modifies the Dockerfile to ensure that a 'node_modules' directory exists prior to copying server files.